Variable renaming is ommited
