+++ findSummary.py
@@ -1,49 +1,60 @@
+#from copyreg import pickle
 import angr
 import sys
+import claripy
+import pickle
 
-def runAndfind(binaryFile):
+class MyConcretizationStrategy(angr.concretization_strategies.SimConcretizationStrategy):
+    def __init__(self, **kwargs): #pylint:disable=redefined-builtin
+        super(MyConcretizationStrategy, self).__init__(**kwargs)
+
+    def _concretize(self, memory, addr, **kwargs):
+        mn,mx = self._range(memory, addr, **kwargs)
+        print("min ",mn," max ", mx)
+        return [mn, mx]
+
+def runAndfind(binaryFile, resultName):
     c = angr.Project(binaryFile, auto_load_libs = False)
     state = c.factory.entry_state()
     #state = c.factory.call_state(0x400671)
+    x = claripy.BVS('x', 32, explicit_name=True)
+    state.memory.read_strategies = [MyConcretizationStrategy()]
+    state.memory.write_strategies = [MyConcretizationStrategy()]
     sm = c.factory.simulation_manager(state)
     sm.explore()
     print(len(sm.deadended))
-    list_list_1 = []
+ 
     if(len(sm.deadended)>0):
+        #fileName = resultName+".txt"
+        f = open(resultName+".txt", "w")
+        f.write(str(len(sm.deadended)))
+        f.close()
         for i in range(len(sm.deadended)):
             path = sm.deadended[i]
             print(len(path.solver.constraints))
-            list_ = path.solver.constraints
+ 
             try:
                 m_1 = path.regs.eax
             except:
                 m_1 = path.regs.r0
-            list_.append(m_1)
-            print(path.solver.constraints)
-            #print(path.regs.eax)
-            list_list_1.append(list_)
-    return list_list_1    
 
+            constraint_and = claripy.And(True)
+            for j in range(len(path.solver.constraints)):
+                constraint_and = claripy.And(path.solver.constraints[j], constraint_and)
+                
+            constraint_and = claripy.And(constraint_and, x==m_1)
+            print(constraint_and)
+            fileName = resultName + str(i)+".pkl"
+            f = open(fileName, "wb")
+            pickle.dump(constraint_and, f)
+            f.close()   
 
-def writeFile(list_list):
-    list_1 = ""
-    for i in range(len(list_list)):
-        if len(list_list[i])==0:
-            continue
-        sublist = ""
-        for j in range(len(list_list[i])):     
-            sublist = sublist + "#"+ str(list_list[i][j])
-        list_1 = list_1 + ","+ sublist 
-    return list_1
     
 
 print(str(sys.argv))
 listName_ = sys.argv
 fileName = listName_[1]
-resultName = str(listName_[2])+".txt"
+resultName = str(listName_[2])
 print(fileName," ",resultName)
-list_list_1 = runAndfind(str(fileName))
-list_list = writeFile(list_list_1)
-f = open(resultName, "w")
-f.write(str(list_list))
-f.close()
+runAndfind(str(fileName), resultName)
+
